Fixed multiple things related to downward chronometers, including adding a "chrono completed" message that flashes the screen. Also fixed the known timestamp, which I forgot to multiply by 1000 (because internal timestamps are not raw unix, they have millisecond support)

+++ src/timeProvider.cpp
@@ -123,7 +123,7 @@ int getRTCisUnadjusted() {
   //   (because that date has passed on all timezones and of course no adjusted clock will have it).
   //this function is only meant to be useful for detecting if the RTC performed a reset, e.g. after taking out the batteries.
   //in that case, the RTC usually resets to some date around 2010. So this function will serve its purpose.
-  if(KNOWN_PAST_TIMESTAMP > currentUnixTime()) {
+  if((long long int)KNOWN_PAST_TIMESTAMP > currentUnixTime()) {
     return 1;
   } return 0;
 }
